QAI-10345: Fix negative std-dev in sfnt-pingpong
The std-dev result for runs of sfnt-pingpong are sometimes negative.
This is caused, in the variance calculation, by 'sumsq' being a
signed 64-bit int, so when diff * diff is calculated this may
overflow in positive 64-bit space. Changing the type of 'sumsq'
and 'variance' to double prevents this (though may sacrifice
precision).

+++ src/sfnt-pingpong.c
@@ -126,7 +126,7 @@ struct stats {
   int64_t median;
   int64_t max;
   int64_t percentile;
-  int64_t stddev;
+  uint64_t stddev;
 };
 
 
@@ -1084,7 +1084,7 @@ static void do_pings(int ss, int read_fd, int write_fd, int msg_size,
 static void get_stats(struct stats* s, int64_t* results, int results_n)
 {
   int64_t* results_end = results + results_n;
-  int64_t variance;
+  double variance;
 
   qsort(results, results_n, sizeof(int64_t), &sfnt_qsort_compare_int64);
   sfnt_iarray_mean_and_limits_int64(results, results_end, &s->mean, &s->min, &s->max);
@@ -1092,7 +1092,7 @@ static void get_stats(struct stats* s, int64_t* results, int results_n)
   s->median = results[results_n >> 1u];
   s->percentile = results[(int64_t) (results_n * cfg_percentile / 100)];
   sfnt_iarray_variance_int64(results, results_end, s->mean, &variance);
-  s->stddev = (int64_t) sqrt((double) variance);
+  s->stddev = (uint64_t) sqrt(variance);
 }
